Bug fix and formatting for legacy reader for FLORIS V2.4 input files (#387)
* Move legacy input reader to separate tool and apply bug fix to work with recent changes in turbine_types

* Add an example to demonstrate legacy input reader

* Remove examples for legacy input reader. Add a __main__ script to the legacy reader file as an example to demonstrate file v2->v3 conversion.
